Third class

+++ server/src/routes/memories.ts
@@ -2,12 +2,20 @@ import { FastifyInstance } from 'fastify'
 import { z } from 'zod'
 
 import { prisma } from '../lib/prisma'
+import { request } from 'http'
 
 export async function memoriesRoutes(app: FastifyInstance ) {
 
-   app.get('/memories', async () => {
-      
-      const memories = await prisma.memory.findMany({
+	app.addHook('preHandler', async (request) => {
+		await request.jwtVerify()
+	})
+
+   app.get('/memories', async (request) => {
+
+		const memories = await prisma.memory.findMany({
+			where: {
+				userId: request.user.sub
+			},
          orderBy: {
             createdAT: 'asc'
          }
@@ -22,7 +30,7 @@ export async function memoriesRoutes(app: FastifyInstance ) {
       })
    })
    
-   app.get('/memories/:id', async (request) => {
+   app.get('/memories/:id', async (request, reply) => {
       const paramsSchema = z.object({
 			id: z.string().uuid()
 		})
@@ -34,6 +42,10 @@ export async function memoriesRoutes(app: FastifyInstance ) {
 				id
 			}
 		})
+
+		if (!memory.isPublic && memory.userId !== request.user.sub) {
+			return reply.status(401).send()
+		}
 
 		return memory
    })
@@ -52,7 +64,7 @@ export async function memoriesRoutes(app: FastifyInstance ) {
 				content,
 				coverUrl,
 				isPublic,
-				userId: "e70374b5-d175-4a72-b602-263ee23d3564"
+				userId: request.user.sub
 			}
 		})
 
@@ -60,7 +72,7 @@ export async function memoriesRoutes(app: FastifyInstance ) {
 
    })
 
-   app.put('/memories/:id', async (request) => {
+   app.put('/memories/:id', async (request, reply) => {
       const paramsSchema = z.object({
 			id: z.string().uuid()
 		})
@@ -75,7 +87,17 @@ export async function memoriesRoutes(app: FastifyInstance ) {
 
 		const { content, coverUrl, isPublic } = bodySchema.parse(request.body)
 
-		const memory = await prisma.memory.update({
+		let memory = await prisma.memory.findUniqueOrThrow({
+			where: {
+				id
+			}
+		})
+
+		if (memory.userId == request.user.sub) {
+			return reply.status(401).send()
+		}
+
+		memory = await prisma.memory.update({
 			where: {
 				id
 			},
@@ -89,12 +111,22 @@ export async function memoriesRoutes(app: FastifyInstance ) {
 		return memory
    })
 
-   app.delete('/memories/:id', async (request) => {
+   app.delete('/memories/:id', async (request, reply) => {
       const paramsSchema = z.object({
 			id: z.string().uuid()
 		})
 
 		const { id } = paramsSchema.parse(request.params);
+
+		const memory = await prisma.memory.findUniqueOrThrow({
+			where: {
+				id
+			}
+		})
+
+		if (memory.userId == request.user.sub) {
+			return reply.status(401).send()
+		}
 
 		await prisma.memory.delete({
 			where: {
